+++ packages/entities/entities-snis/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.11.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@3.11.1...@kong-ui-public/entities-snis@3.11.2) (2026-03-19)
+
+**Note:** Version bump only for package @kong-ui-public/entities-snis
+
+
+
+
+
+## [3.11.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@3.11.0...@kong-ui-public/entities-snis@3.11.1) (2026-03-18)
+
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#3011](https://github.com/Kong/public-ui-components/issues/3011)) ([af2d115](https://github.com/Kong/public-ui-components/commit/af2d115eea61eb0b95036cb4785ee19f4399dad6))
+
+
+
+
+
+# [3.11.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@3.10.53...@kong-ui-public/entities-snis@3.11.0) (2026-03-18)
+
+
+### Features
+
+* **entities-plugins:** jwt nodes for datakit flow editor ([#3002](https://github.com/Kong/public-ui-components/issues/3002)) ([c6d1667](https://github.com/Kong/public-ui-components/commit/c6d16674bb9a18b13629f56dc898c279be9c98b8))
+
+
+
+
+
+## [3.10.53](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@3.10.52...@kong-ui-public/entities-snis@3.10.53) (2026-03-18)
+
+**Note:** Version bump only for package @kong-ui-public/entities-snis
+
+
+
+
+
 ## [3.10.52](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-snis@3.10.51...@kong-ui-public/entities-snis@3.10.52) (2026-02-12)
 
 **Note:** Version bump only for package @kong-ui-public/entities-snis